+++ src/main/java/com/asociatialocatari/gal/apart/ApartDto.java
@@ -2,6 +2,8 @@ package com.asociatialocatari.gal.apart;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ApartDto {
 
@@ -15,7 +17,7 @@ public class ApartDto {
 
     private Integer tenantsNumber;
 
-    private Double m2;
+    private BigDecimal m2;
 
     private Long idBuildStair;
